Color coding, keyword search of titles
These changes to the RSS stream provide color coded titles based on sentiment analysis of title keywords.  Positive and negative keywords and their associated scores are in two files in the project root directory.  Positive colors are dim green and bright green.  Negative colors are dim red and bright red.  Instead of dumping json to stdout, the data for an article is printed to the screen.